osd: Make shard_id_t stricter and deal with the fallout.

+++ src/common/scrub_types.cc
@@ -117,7 +117,7 @@ void inconsistent_obj_wrapper::add_shard(const pg_shard_t& pgs,
                                          const shard_info_wrapper& shard)
 {
   union_shards.errors |= shard.errors;
-  shards.emplace(osd_shard_t{pgs.osd, int8_t(pgs.shard)}, shard);
+  shards.emplace(osd_shard_t{pgs.osd, pgs.shard}, shard);
 }
 
 void
